requirements.txt
pytest==7.4.3
selenium==4.15.2
allure-pytest==2.14.2
pytest-selenium==4.1.0 